Set test environment
