+++ se_project.iml
@@ -4,7 +4,7 @@
     <exclude-output />
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/unit_test" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/src/test" isTestSource="true" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
@@ -20,5 +20,6 @@
     </orderEntry>
     <orderEntry type="library" name="org.testfx:testfx-junit:4.0.13-alpha" level="project" />
     <orderEntry type="library" name="org.testfx:testfx-core:4.0.13-alpha" level="project" />
+    <orderEntry type="library" name="org.loadui:testFx:3.1.2" level="project" />
   </component>
 </module>
