+++ src/view/Main.java
@@ -1,3 +1,5 @@
+package view;
+
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
